pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codexs-ai-bot"
version = "0.1.0"
description = "Codexs.ai bilingual Telegram assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==21.5",
    "python-dotenv==1.0.1",
    "httpx==0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
]
authors = [
    { name = "Codexs.ai" },
]

[project.scripts]
codexs-bot = "codexs_bot.bot:main"

[tool.setuptools.packages.find]
where = ["src"]


